BIGUpdate

+++ src/main/java/com/project/JewelryMS/service/PerformanceService.java
@@ -45,7 +45,7 @@ public class PerformanceService {
 //        return performanceRepository.save(performance);
 
         //the staffAccountRepository could have a JpaRepository of <Integer>
-        Optional<StaffAccount> staffOptional = staffAccountRepository.findById(createPerformanceRequest.getStaffID());
+        Optional<StaffAccount> staffOptional = staffAccountRepository.findById((long)createPerformanceRequest.getStaffID());
         if(staffOptional.isPresent()) {
             StaffAccount account = staffOptional.get();
             if (account != null) {//This if condition needs revision, should use to connect with Shift to accurately create a performance report
